Add several backing tracks and randomly choose one for song
